Stop the code if all the calculations are done

+++ src/geode.h
@@ -88,9 +88,9 @@ extern void vis(GLuint, size_t);
 #endif
 
 // Geode specific functions
-extern void  dump  (Data &);
-extern float evolve(Data &, double);
-extern void  init  (Data &);
-extern int   solve (Data &);
+extern void   dump  (Data &);
+extern double evolve(Data &, double);
+extern void   init  (Data &);
+extern int    solve (Data &);
 
 #endif // GEODE_H
